refactor(tests): migrate test files to feature-descriptive naming convention
Rename test files to follow TD-1/TD-2/TD-3 conventions:
- auth: e2e_tests → e2e/test_register_login, jwt_tests → unit/test_jwt
- clusters: e2e_tests → e2e/test_cluster_crud, serializer_tests → unit/test_serializer
- deployments: e2e_tests → e2e/test_deployment_crud, serializer_tests → unit/test_serializer
- cross-app: e2e_tests → e2e/clusters_deployments/test_deployment_with_cluster
