+++ app/src/apps/auth/tests/e2e/test_register_login.rs
