+++ app/src/apps/deployments/tests.rs
@@ -1,4 +1,8 @@
 //! Tests for deployments app.
 
-pub mod e2e_tests;
-pub mod serializer_tests;
+pub mod e2e {
+	pub mod test_deployment_crud;
+}
+pub mod unit {
+	pub mod test_serializer;
+}
