+++ app/tests/e2e.rs
@@ -1,0 +1,1 @@
+mod clusters_deployments;
