+++ app/tests/e2e/clusters_deployments.rs
@@ -1,0 +1,1 @@
+mod test_deployment_with_cluster;
